Fixed opennaas-core-persistence.
It is important persistence-api is at least 2.0 or we'll get javax.persistence.OneToMany.orphanRemoval().
dynamic-jpa from org.dynamicjava.osgi enforces 1.0.2, which is troublesome old garbage. We've disabled that dep for now, things seem to just work.
It seems that order matters and jpa-api should be as high as possible on the pom.