resources: Fix ClassCastException on cloning ResourceDescriptor
When cloning a resource descriptor that was loaded from persistent
storage, the following ClassCastException is thrown:

Caused by: java.lang.ClassCastException: org.hibernate.collection.PersistentMap cannot be cast to java.util.HashMap
	at org.opennaas.core.resources.descriptor.ResourceDescriptor.clone(ResourceDescriptor.java:172)
	at net.i2cat.mantychore.queuemanager.QueueManager.newResourceDescriptor(QueueManager.java:276)
	... 28 more

This is caused by the code assuming that the properties map is allways
a HashMap. For a ResourceDescriptor loaded from persistent storage this
assumption is not true.

The patch replaces the attempt to clone the properties map by a simple
copy of the collection to a newly allocated HashMap.