persistence: Fix filter for EntityManagerFactory service lookup
GenerociOSGIJpaRepository asks OSGi for an EntityManagerFactory
service.

Such services are offered by Dynamic JPA, a bundle that detects
persistence.xml files and automatically creates EntityManagerFactory
services for each persistence unit. DynamicJPA exposes two
EntityManagerFactory services for each persistence unit: The real
EntityManagerFactory and a wrapper that hides restarts and changes of
the bundle providing the persistence unit. To disinguish the two,
DynamicJPA adds the isDynamicFactory property to the service.

GenericOSGIJpaRepository attempts to filter the service selection using
this property, but as the log file shows the filter is not as we
expect:

 01-19@16:23:00 [Blueprint Extender: 3]DEBUG org.opennaas.core.persistence.GenericOSGiJpaRepository - Looking up Service from registry with properties: (&(objectClass=javax.persistence.EntityManagerFactory)(isDynamicFactory=null)(persistenceUnit=ResourceCore))

Notice the check isDynamicFactory=null when we would have expected
isDynamicFactory=true.

The problem is that the Properties instance used by
GenericOSGiJpaRepository is used in a type-unsafe manner. The
Properties class inherits from Hashtable, but should only be used for
String key and String values. The methods of Properties provide
type-safety, but the methods inherited from Hashtable do not. Thus the
compiler doesn't complain when attempting to use a boolean as a
property value rather than a string. The result is that getProperty
returns null later on.

This patch replaces the boolean with a string value and changes the
code to call the type-safe setProperty rather than put.