Use type-safe API of Properties
Properties only opertors on String key and String values. The methods
inherited from Hashtable should be avoided as they are not type-safe.

This patch updates the two createServiceFilter methods to use the type
safe stringPropertyNames() method rather than the keys() method. The
stringPropertyNames() method also takes chained Properties instances
into account - something the keys() method doesn't.

The patch changes the use of the legacy Enumerator class to the use of
a for iteration.

Finally the patch updates both methods to avoid the repeated string
concatenation in a loop. It does so by using a StringBuilder (this is
not strictly related to the rest of the patch, but is generally
considered best practice).