resource,persistence: Make ResourceDescriptorRepository closeable
ResourceDescriptorRepository (or rather its base class
GenericOSGiJpaRepository) creates EntityManager instances from an
EntityManagerFactory. Unfortunately these EntityManagers are never
closed, leading to resources leaks and persistence problems when the
bundle is stopped.

This patch makes two changes:

- It adds a close method to GenericOSGiJpaRepository to give it a chance
  to close the EntityManager.

- If changes ResourceRepository such that the ResourceDescriptorRepository
  is injected rather than created internally.

The latter change allows the container to manager the lifecycle of the
ResourceDescriptorRepository (and thus call close).

Unfortunately the latter change requires changes in plugins as they
need to instantiate the ResourceDescriptorRepository and inject it into
the ResourceRepository.

I considered to add a close method to ResourceRepository instead,
however ResourceRepository already allows the dependency to be injected
(it can operate with either an injected GenericRepository or an
internally created ResourceDescriptorRepository). Thus one would have
to track whether the repository is created internally or injected and
only close it when created internally. I concluded that this is too
messy and it would be cleaner to always inject the dependency. That
this is the correct solution is also demonstrated by unit tests
currently having to set an unused persistence unit when calling the
constructor. In the refactored code this is no longer necessary.

Matching changes for Mantychore will be emailed to Isart.