resources: Shutdown HSQLDB automatically
HSQLDB needs to be shutdown before terminating the application.
Otherwise unwritten (but otherwise committed) changes will not be
written to disk. Since the delayed write interval is configured to 10
seconds, this is quite likely to result in dataloss when resources are
modified just prior to shutting down OpenNaaS.

The normal way to shutdown HSQLDB is to execute the "shutdown" SQL
command. This does however bind the application code to HSQLDB and I
find this undesirable for the production code. The alternative way is
to set the shutdown=true property. This triggers an automatic shutdown
when the last connection is closed.

Unfortunately this change by itself is not enough to cause HSQLDB to
shut down. This is because the Dynamic JPA bundle never closes its
EntityManagerFactory and thus the last reference to HSQLDB is never
closed.

This patch adds the shutdown=true for the production persistence unit;
the test persistence unit is not modified.