jpa: Replace Dynamic JPA with Aries JPA
Dynamic JPA contains a bug in which it never closes the
EntityManagerFactory instances it creates. This is a particular problem
with HSQLDB which relies on the EntityManagerFactory being closed to
trigger a shutdown of the embedded database.

It also seems the Dynamic JPA is no longer maintained, making it a bad
choice. Finally, Dynamic JPA doesn't respect the OSGi Enterprise
Specification (which is possibly why the bundle is no longer
maintained).

OpenNaaS already depends on Apache Aries and Aries happens to have JPA
support. This patch therefore replacces Dynamic JPA with Apache Aries
JPA.

A number of changes had to be made to the Hibernate packaging: The
Hibernate package now exports a PersistenceProvider service which
allows it to be picked up by Aries (this follows the OSGi Enterprise
spec). I had to remove the packaging of javax.persistence to avoid
conflicts. I also replaced the fixed import of org.hsqldb with a
dynamic import to allow alternative JDBC drivers to be used (eg move
the database to PostgreSQL for a more robust database backend).

An extra metadata header had to be added to the resources bundle to
allow persistence.xml to be picked up.

Finally, the jpa feature needs to be added as a boot feature. The same
change needs to be done to Mantychore and other tools building on
OpenNaaS.