Use type-safe API of Properties.
Properties only operates on String key and String values. The methods
inherited from Hashtable should be avoided as they are not type-safe.

This patch replaces the use of properties.put(String, Object) method passing a String as value, to use the type safe setProperty(String, String) one.

The patch doesn't fix anything and is only for consistency with the rest 
of the code (and to guard against future mistakes).

git-svn-id: http://svn.i2cat.net/repos/manticore/trunk@1755 e5892e71-24e5-41bb-b533-28a584b410ba