packaging: Upgrade to latest release of FUSE ESB
The patch upgrades OpenNaaS to the latest release of FUSE ESB. While
doing so it attempts to parameterize references to version numbers of
FUSE components.

This release of FUSE changes how features installed through the
bootFeatures setting are loaded. This solves some problems in which
features that would install normally after startup would fail when
installed with other features through the bootFeatures property.
However as a consequence features now need to be self-contained. Ie it
is no longer suffient to refer to cxf and activemq features in
bootFeatures; the opennaas features have to depend on those features
directly.

To add the necessary dependencies to the opennnaas features I had to
switch the features-maven-plugin from org.apache.felix.karaf.tooling to
the newer org.apache.karaf.tooling version.

Finally, I changed the featuresBoot list to be based on the new FUSE
default, with opennaas features added. This solves the problem of
the console service being restarted and a stack traces printed the
first time OpenNaaS is started.