Remove opennaas unrelated examples from platform.
As stated in issue
http://jira.i2cat.net:8080/browse/OPENNAAS-180

OpenNaaS platform contains a folder with irrelevant examples.
These examples come from service-mix. They are not related to opennaas and should, be removed.

They are added to platform target when service-mix is unpacked as part of platform building.
Service-mix is unpacked within a maven-dependency-plugin execution with id unpack-apache-servicemix on package phase.

This patch configures the above mentioned execution for excluding examples folder.