tests: Let net.i2cat.nexus.resources.tests use injection
The patch cleans up the net.i2cat.nexus.resources.tests (currently
the only tests using the KarafTestContainer) such that dependency
injection is used to obtain references to other services.

The patch also removes the polling code and delay to wait for
other services. The intention with that code was to give the
blueprint containers time to start and thus ensure that
repositories and capabilities needed for the tests had been
loaded. It is faster and more robust to instead inject the
blueprint container reference into the test. This ensures that
the test isn't executed until the required blueprint container
has finished initilization. The added benefit is to make the
dependency explicit in the code.

Please note that I do not guarantee that I recognized all
dependencies. If you see the tests fail non-deterministically
then please tell me so we can add all dependencies.