tests: Port weeb26 tests to pax exam 2
Note that the tests are not actually included in OpenNaaS/Manticore
(they are commented out of manticore/tests/sprints/pom.xml). I assume
this is because the tests are failing.

One of the ported tests still fails (InterfacesVLANKarafTest). As
far as I can see this is because the test is buggy. It fails with:

[ERROR] the command encapsulation is limited in logical tunnels