Updated RoutingOptionsParser to parse RouterId.
This patch allows the parsing of router-id field.
It also changes the behavior of RoutingOptionsParser:
It now requires a model as a parameter (in the constructor) and updates given model with routerId and NextHopRoutes.
The fact of routerId being stored in RouteCalculationServices, and those Services not being populated by RoutingOptionsParser, requires this parser to take the model as a parameter.