Updated protocols parser to parse OSPF configuration.
ProtocolsParser contains rules to populate router model with OSPF information gathered from the router configuration in xml.

A router having OSPF has an OSPFService instance available using HostedServices association.
OSPFService stores the routerID. It also stores enables/disabled state.

OSPFService contains an OSPFAreaConfiguration for each OSPFArea.
OSPFService is associated to its OSPFAreaConfiguration through OSPFServiceConfiguration association.
AreaConfiguration and Area are associated through AreaOfConfiguration.
Each area is identified by its areaID. This id is a long in the model, and should be formatted from the dotted ip address format used in the xml.

Each OSPFArea contain several OSPFProtocolEndpoints tied to NetworkPorts of the model.
OSPFArea is associated to its protocol endpoints using EndpointInArea.
An OSPFProtocol is tied to its NetworkPort through PortImplementsEndpoint association.
Each OSPFProtocolEndpoint is marked as enables/disables according to the xml.

We'd like to add a dependency between an OSPFProtocolEndpoint and an IPProtocolEndpoint, meaning that the latter is required for the lesser.
However, Junos associates OSPF areas to interfaces, not to IP addresses. And adding a dependency for each IP address in that interface is not what we want.
We still have to find a way to specify the constraint of OSPFProtocolEndpoint requiring an IPProtocolEndpoint in associated interface.