build: Don't append assembly id to artifact
A recent change upgraded the assembly plugin. The new version
required that all assembly descriptors have an id. By default
the id is appended to the artifact, which meant the opennaas
assembly artifact change name. This change was not intended and
broke the assembly of manticore.

This patch changes the assembly pom such that the assembly id
is not appended to the artifact name, thus restoring the original
name.