Modified ConfigureOSPFInterfaceStatusAction:
- It searches into model required OSPFArea information.
- Generated template contains each area once only.

ConfigureOSPFInterfaceStatusAction needs to know the OSPFArea ID each endpoint belongs to in order to craft the rpc message.
OSPFArea ID is gathered from the model. An exception is thrown if this informaion is not available.

Velocity template has changed to accept an OSPFService as a parameter, and it now creates an rpc message with each area appearing only once.

I had to update some tests, also.