tests: Fix roadm repository dependency
The test mistakenly injected the roadm repository itself rather
than the blueprint service. So even though the test would
wait for the repository service to be instantiated, it would not
wait for the end of the blueprint initialization and thus
registration with the repository manager. This could occasionally
lead to test failures.