tests: Relocate net.i2cat.nexus.tests.helper
This module doesn't contain tests, it contains utility classes
used by tests. Placing the module on the tests module makes it
difficult to move common dependencies for tests into the parent
poms. If we did then the helper module would be affected too
even though it isn't a test. Eg we don't want the helper to
be subject to pax exam execution.

This patch therefore moves this module to the manticore/bundles
module. The patch also deletes the now unused IntrationTestsHelper
class.

The patch will make future patches for cleaning up the poms
of test modules simpler.