tests: Avoid sleeping in tests
Quite a number of test use sleeping as a form of synchronization.
Besides being bad style and fragile, this adds to the runtime
of the tests.

This patch removes some of the sleeps and replaces them with other
means of synchronization.

There are tests left that sleep, but they are not easily fixed and
require deeper changes.

Tests done: Unit and integration tests pass.