build: Configure Maven release plugin
The Maven release plugin helps with branching, tagging,
and version number increments. The plugin greately
simplifies building a release and should always be used
when making releases.

The release plugin is configured to perform branching,
tagging and commits in the local git repository. The
changes have to be pushed upstream manually afterwards.
This also means it is fairly safe to experiment with
the release plugin as the changes can easily be rolled
back.