build: Configure site module
In Maven 3 the site module needs to be configured to allow for
reporting plugins to work. This patch configures the site plugin.

The patch also configures various reports, including JavaDoc, source
code cross reference, find bugs analysis, pmd and cpd analysis.

I disabled a couple of slow reports, and generally chose aggregated
reports over individual reports.

I had to update the two platform projects to ensure that the
features-magen-plugin is executed in the correct phase (otherwise it
would conflict with the JavaDoc plugin).

The site report can be build by running:

    mvn clean install site

followed by

    mvn site:stage

The latter combines all reports into a single report structure in
target/staging/

Note that producing the reports will take a lot of memory and the
default Maven settings will not suffice. I set

    MAVEN_OPTS="-Xmx2048m -XX:MaxPermSize=1024M"

to solve this.