build: Rename features to opennaas naming scheme
The patch adjusts the feature names to avoid the name 'mantychore'. All
feature names are prefixed opennaas- (with a couple of exceptions for
test bundles).

I am not using the proposed opennaas-extensions- prefix for feature
names as it would make the names unusually long compared to other
ServiceMix feature names. Prefixing with opennaas- should be enough to
make the names fairly unique.

I renamed mantychore-base to mantychore-cim as the feature is about
representing the CIM model. In its current form the feature also holds
the network model, but this is because the dependency between the two is
upside down. Once the dependency has been inverted the network model
bundle should be moved to the network feature. Similarly the
queuemanager bundle should be part of opennaas-core, but the bundle
needs to be renamed and moved first.

The opennaas-luminis feature should possibly be split and renamed, but
for now I have put those bundles into a single feature.