scripts: Workaround for NoSuchElementException in Karaf
The patch addresses issue opennaas-162. The problem is that
Karaf throws a NoSuchElementException when entering an
empty command while the Karaf history is empty (ie on first
run until a command has been exexcuted).

The bug is fixed in newer releases of Karaf, however the
fusesource release does not contain the fix yet.

This patch adds a simple workaround: The mantychore.sh script
checks for a missing karaf history file. If no history file
is found, a file with a single empty command is created.

The workaround should be removed once the upstream fix is
merged into the fusesource release.