platform: Merged manticore/platform into platform
We used to have both an OpenNaaS platform project without extensions and
a Mantychore platform project including extensions.

This patch merges these such that we only got one platform project. This
also means that the feature repositories get merged and we only need one
assembly project.

I added all artifacts as dependencies in the platform project. This is
to ensure that Maven builds the artficats in the correct order. As a
consequence I had to move the dependencyManagement section of
manticore/pom.xml to pom.xml.