build: Refactor platform assembly
The packaging step was split into two Maven projects earlier:
the platform and the assembly. This patch refactors this into
a single Maven project.

We now rely more on the assembly plugin and less on the
dependency and resource plugins to drive the assembly. I find
this makes it a bit easier to understand what is going on.

The resulting artifacts are now stored in platform/target/.

Note that the actual project version number is now used to
name the packages.