Renamed extensions artifacts
They now follow the naming convention (ie org.opennaas+path
elements). The exception to the rule is that the bundles folder
name is not included in the artifact ID as I expect that folder
to be deleted.

Java package named are unaltered. Those will be renamed later.

Defaults for bundle symbolic name are now defined in the top
level pom and should not be repeated in subprojects.