Created velocity templates to add/remove tagged ethernet encapsulation and set vlan-ids
Tagged ethernet encapsulation is set in junos config in a different way deppendeing on the interface type.
This encapsulation is allowed only in physical eth interfaces and in logical lt interfaces.
- In physical eth interfaces, a <vlan-tagging> tag is used
- In logical lt interfaces, an <encapsulation> element with "vlan" value is used. It has to be noticed that
this <encapsulation> element is the same used to specify plain ethernet encapsulation (using "ethernet" value).

VlanId can be set only in logical interfaces.
A <vlan-id> element with desired vlanId value is used.