Added checks to setEncapsulation actions.
Note: Checking an interface exists in the router is not yet possible.
Right now, we only can check interface exists in the candidate, but that's not what we want:

An interface is removed from the configuration when it has nothing configured.
That is a typical case when setting encapsulation, as interface may have no encapsulation nor units.
Unless this interface contains further configuration (description, or whatever else) it will not appear in candidate.

For this reason, setEncapsulation actions do not check an interface exists or not,
 and when executed, they create such interface in the configuration.