Refactored ProtocolSessionManager.
A part from an interal refactor and reordering of methods,
a method in the interface has been renamed.

It happened that there were 3 methods called obtainSession*:
- two of them tryed to get an existing session and created a new one if there was none,
- but the other one (obtainSessionById) returned an exception when failed to get en existing session.
This third method has been renamed to getSessionById, with no functional changes, to avoid confusion.

Additionally, some exception messages have been changed.

The only functional change is the assumption that there is only one protocolSession per registered context.
This assumption is taken into account when unregistering a context, as we stop looking for protocolSessions associated with it when we find the first one.