Include ProtocolSessionManager creation/destruction in resource lifecycle.
Having them decoupled, may lead to sessions not being closed until opennaas is closed,
as described in OPENNAAS-255 issue (http://jira.i2cat.net:8080/browse/OPENNAAS-255)

This patch causes ResourceRepository to create a ProtocolSessionManager when a resource is created,
and destroy it when the resource is removed.

Sessions are destroyed when destroying the manager, due to previous commits.