Update ResourceRepository to create and destroy sessions when resource is created/removed.
ResourceRepository is responsible of synchronizing resource lifecycle with protocolSessionManager one.
- It creates a protocolSessionManager for the resource when the resource is created.
- It destroys the protocolSessionManager for a resource, when the resourceis removed.
- It also creates one session per registered context when a resource is started, if it does not already exists.
 (This assures a single session per each registered context is available after a starting the resource.)
- It also destroys all sessions of a resource when the resource is stopped.
 (This assures no session is available after stopping the resource.)

This patch adds session creation/deletion when resource is started/stopped.

It also fixes a bug when rolling back from deactivateResource.
The bug caused bootstrapper and capabilities to be reseted even if deactivate failed, producing an unusable resource.
Patched code makes this reset only when deactivate is succesfull.