autobahn: Add transaction support
I found a solution to the problem of attaching cross-action
state during queue execution: A thread local variable holds
the log. As long as all actions on a queue are executed
from the same thread, this solution will work nicely.

Callouts to Autobahn are now performed by commands queued
for execution during confirm. In case of restore the commands
know how to undo what they did.

The pattern I follow here sort of replicates some of the
functionality of QueueManager: Queuing, execution and rollback
of commands is not unlike the queueing, execution and rollback
of actions done by QueueManager. I get the feeling that
the Action concept could be improved such that the execute and
undo actions could be part of the actions themselves, thus
avoiding the need to implement something like the Transaction
class in the driver.