autobahn: Added Response items to ActionResponses
Action responses can contain a list of Response objects. These
document the actions performed on behalf of the action and their
result. This is convenient now that the bulk of the processing
happens in the ConfirmAction and the RestoreAction.

It is also convenient to use this to report on discovered ports
in GetTopologyAction.

This patch refactors the transaction code of the Autobahn driver
to generate Response objects.

Unfortunately the OpenNaaS QueueManager execute output only
contains this information in case of failures. In case of success
the QueueManager replaces the ActionResponse with an empty one.