autobahn: Reject requests that do not originate at local interfaces
Autobahn can only establish connections from a domain local interface to
some other interface (which may be local to the Autobahn domain, or in
the domain of some other BoD system).

This patch collects information about which interfaces are local to a
particular Autobahn BoD resource and reject requests for which the first
interface is not local.