Break cyclic references when marshalling model to XML.
There was a cyclic reference between ManagedElements in an Association.
To break it, associations are marshalled as unidirectional (using @Transient annotation in one direction).

To make it bidirectional when unmarshalling, afterUnmarshall method is implemented in ManagedElement.
This method is called by JAXB after a ManagedElement is unmarshalled and it populates back Association's not marshalled direction.