Check ResourceDescriptor and ResourceIdentifier marshall/unmarshall works for ws.
Marshalled ResourceIdentifier needs to store the resourceId, so @XmlTransient annotation has been removed.
This is required in order to export IResourceManager through web services.

Just as it was before, Resource Id is stored in two places: ResourceDescriptor.id and ResourceIdentifier.id
This is a duplication of information, but as both ids are used it is not easy to remove one of them.