Modify ProtocolManagerTest by adding extra test for getting sessions when more than one context is used: -if the contexts have different protocols, both contexts are allowed and 2 sessions needed -if the contexts have the same protocol, only one context is allowed (the last registered) and only one session.
The need for extra tests has arisen when adding Key authentication was
incorrectly creating many sessions with the same context as it was
considering there were many context. The solution was rewriting the
ProtocolSessionContext equals operator ( as it was not working correctly when "null"
was being used for KEY_PATH and KEY_PASSPHRASE when AUTH_TYPE was
password).