Publish PSM using REST and DOSGI.
Not all methods of IProtocolSessionManager are exported,
only the ones for managing contexts, and for requesting session
information and deleting sessions.

A dedicated ListResponse object has been necesary to return
List<String>.

Notice that after this patch, each protocolSessionmanager is
published as an osgi service, which was not happening before.