Protocols:context command interactively asks for passwords.
This way passwords don't need to be in the command, which caused
them to be visible and even stored in history file.

Although strongly discouraged, passwords can be specified in the command, just like before.
However, a new way to specify passwords is introduced:

protocols:context -i (--interactive) tells the command to ask for passwords when required.
This way, CLI is used to introduce passwords, which are not displayed nor stored in visible places.