Add FIXME description in CLIProtocolSession.
Prompt values are used as delimiters for cli output.
Hence, having a delimiter used in the output will cause the output to be cut.

Delimiters shoud be specified per resource type, or per session in its
ProtocolSessionContext.

Delimiters currently harcoded in CLIProtocolSession correspond to
cisco catalyst prompt, but should not be used in CLI sesions to other
resources.