Update l2BoDCapability and its drivers to identify links by RequestParameters.
Now l2BoDCapability uses BoDLinks having a property of ConnectionRequestParameter type.

In order to create a link, l2BoDCapability receives a ConnectionRequestParameter object. (This has not changed).
When a link is created, it is associated with its ConnectionRequestParameter.

In order to deprovision this link, l2BoDCapability.shutdownConnection accepts a ConnectionRequestParameter.
This easies clients interaction with the capability.

There is also a l2BoDCapability.shutdownConnection method accepting the Link itself.
This allows a user inspecting the model to select which link to deprovision and easily call l2BoDCapability.

Autobahn driver and dummy driver have been updated to follow this behaviour.
Web services and tnc-demo-web have been updated too, to call new l2BoDCapability interface.

BoD integration tests have been also updated, and now pass.