Check ethernet-tagging encapsulation only in ethernet interfaces.
There is a check preventing the creation of subinterfaces with unit different than 0
when no ethernet-tagging is specified in the physical interface.

This check was introduced when solving issue:
http://jira.i2cat.net:8080/browse/OPENNAAS-338

In pull request
https://github.com/dana-i2cat/opennaas/pull/278

More precisely on commits:
8911c1bf0d8f2c2fcf3597b88cbba3d190b81a6f
6b502d7e63df5a725de1b28e8a9d89234817af3f

Some interfaces may have subinterfaces without requiring ethernet-tagging.
It is the case of logical interfaces, such as the ones used in gre tunnels.

This patch changes the check to be performed only in ethernet interfaces.
This patch allows the following command to work:
chassis:createsubinterface router:name gr-1/2/0.1