Fix bug in autobahn driver useing same vlans in all requests with same interface.
The bug was caused by using same PortType in all requests and, worst, also setting desired vlan in this object for each request.
This caused last request vlan (the last set) being used in all requests and maintained in the model.

This patch solves the problem by creating a copy of the original PortType in each request. Also, vlan is set only in the copy
(not in the original one), maintaining original PortType unchanged.