Create model in VCPEManager with PhysicalTopology.
VCPEManager needs a huge refactor to become a Resource.
PhysicalTopology should also change to be more flexibility.
In fact, it needs same flexibility as a NetworkModel, so using a NetworkResource to model it should be the way to go.

This patch creates the model and makes it accessible through WS.
PhysicalTopology is right now created using a properties file.