VCPE vlan and unit suggestion does not repeat vlans and units in same vcpe.
Suggestion is now improved in a way it does not repeat same vlan for same physical interface.
It does so for units, too.

In order to do that, we now keep track of already suggested vlans and units in current vcpe,
and suggest a particular one only if it has not been suggested already (in the same vcpe).

Although not introduced in this patch, it's worth notice that suggestion also calls isFreeVLAN for a candidate vlan
in order to check the vlan is not being used in other VCPEs.